🔥 (zsh): some simplifications in config
This commit removes the use of <Esc> to clear the screen, because
with the introduction of the magic caps-lock key (with Karabiner),
it happens that I press <Esc> unintentionally, so this removes the
effect of the key in the shell (I can still use `cl` for `clear`,
though).

It also removes the `KEYTIMEOUT` setting, as I don't think it's
useful anymore.